fix folder deletion issue

+++ lib/ModlessManager.as
@@ -115,7 +115,9 @@ namespace ModlessManager {
         sleep(4000); // to be sure the textures were indeed loaded
         auto app = cast<CTrackMania>(GetApp());
         if(ModWorkManager::IsModWorkBTMOnly() && hasTriggeredModWorkFolderCreation) { // If folder existed previously, it means there was custom modwork already and we don't need to load the map twice
-            IO::DeleteFolder(MODWORK_FOLDER, true);
+            if(IO::FolderExists(MODWORK_FOLDER)) {
+                IO::DeleteFolder(MODWORK_FOLDER, true);
+            }
             if(app.RootMap.ModPackDesc !is null) { // we only need to reload twice if map has a mod
                 MapLoading::ReloadMap();
             }
