+++ lib/settings/AdvancedSettings.as
@@ -41,8 +41,12 @@ namespace AdvancedSettings {
             
             if(ConfirmationButton(Icons::Exchange + " Switch to Modless", vec2 ( ), "Switching to Modless requires to delete the ModWork folder, are you sure ?")) {
                 modMethod = "Modless";
-                IO::DeleteFolder(MODWORK_FOLDER, true);
-                IO::DeleteFolder(MODWORK_DISABLED_FOLDER, true);
+                if(IO::FolderExists(MODWORK_FOLDER)) {
+                    IO::DeleteFolder(MODWORK_FOLDER, true);
+                }
+                if(IO::FolderExists(MODWORK_DISABLED_FOLDER)) {
+                    IO::DeleteFolder(MODWORK_DISABLED_FOLDER, true);
+                }
                 UI::ShowNotification(Icons::Kenney::TimesCircle + " Better Texture Mod - ModWork folder deleted", "ModWork folder has been deleted and plugin will be Modless after restart.", UI::HSV(0.51, 0.69, 0.9), 8000);       
             }
             UI::TextWrapped("Modless method allows you to have plugin textures alongside custom mods. However you will not be able to use new wood/old wood texures presets. Keep in mind any texture of a custom mod will override its corresponding texture in the plugin !\n"
@@ -89,8 +93,12 @@ namespace AdvancedSettings {
         if(IO::FolderExists(MODWORK_FOLDER)) {
             UI::PushStyleColor(UI::Col::Button, vec4(1,0,0,1));
             if(ConfirmationButton(Icons::TrashO + " Delete ModWork folder", vec2 ( ), "Are you sure you want to delete the ModWork folder ?")) {
-                IO::DeleteFolder(MODWORK_FOLDER, true);
-                IO::DeleteFolder(MODWORK_DISABLED_FOLDER, true);
+                if(IO::FolderExists(MODWORK_FOLDER)) {
+                    IO::DeleteFolder(MODWORK_FOLDER, true);
+                }
+                if(IO::FolderExists(MODWORK_DISABLED_FOLDER)) {
+                    IO::DeleteFolder(MODWORK_DISABLED_FOLDER, true);
+                }
                 UI::ShowNotification(Icons::Kenney::TimesCircle + " Better Texture Mod - ModWork folder deleted", "ModWork folder has been deleted.", UI::HSV(0.51, 0.69, 0.9), 8000);       
             }
             UI::PopStyleColor(1);
